proto1/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.51.2
python-dotenv==1.0.0
pydantic==2.11.3
python-multipart==0.0.6